+++ app/controllers/products_controller.rb
@@ -16,25 +16,37 @@ class ProductsController < ApplicationController
   end
 
   def update
-    if @product.user_id == @current_user.id
-      @product.update_attributes product_params 
-      save_and_render @product
-    else
-      permissions_error
+    if product_does_not_have_purchases?
+      if @product.user_id == @current_user.id
+        @product.update_attributes product_params 
+        save_and_render @product
+      else
+        permissions_error
+      end
     end
   end
 
   def destroy
-    if @product.user_id == @current_user.id
-      render_ok @product.destroy
-    elsif is_current_user_admin.nil?
-      render_ok @product.destroy
-    end
+    if product_does_not_have_purchases?
+      if @product.user_id == @current_user.id
+        render_ok @product.destroy
+      elsif is_current_user_admin.nil?
+        render_ok @product.destroy
+      end
+   end
   end
 
   private 
   def set_product
     @product = Product.find params[:id]
+  end
+
+  def product_does_not_have_purchases?
+    if @product.purchases.empty? 
+      true 
+    else  
+      render json: {authorization: 'You can not edit/destroy products that users already bought, we have to preserve the history'}, status: :unprocessable_entity
+    end
   end
 
   def product_params
